requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyterlab>=4.0.0
pyyaml
requests
beautifulsoup4
wordcloud
UzMorphAnalyser